refactor: move abi.rs to zone-primitives crate
Move the Solidity ABI bindings from the zone crate into zone-primitives
so they are available to no_std consumers (zone-stf, SP1 guest) without
pulling in the full zone crate.

A `define_abi!` macro conditionally injects `#[sol(rpc)]` on each
contract when the `rpc` feature is active, keeping the heavy
alloy-contract dependency optional. The `serde` feature is also gated
so bare no_std builds stay lean.

The zone crate's abi.rs becomes a thin re-export (`pub use
zone_primitives::abi::*`) plus a `ZonePortalExt` trait for the async
RPC helper methods that require alloy-provider and tokio.
